+++ lib/msal-common/src/authority/AuthorityOptions.ts
@@ -32,9 +32,6 @@ export const AzureCloudInstance = {
     // Microsoft Azure public cloud
     AzurePublic: "https://login.microsoftonline.com",
 
-    // Microsoft PPE
-    AzurePpe: "https://login.windows-ppe.net",
-
     // Microsoft Chinese national/regional cloud
     AzureChina: "https://login.chinacloudapi.cn",