+++ samples/msal-angular-samples/angular-standalone-sample/src/environments/environment.e2e.ts
@@ -3,11 +3,11 @@ export const environment = {
   msalConfig: {
     auth: {
       clientId: '3fba556e-5d4a-48e3-8e1a-fd57c12cb82e',
-      authority: 'https://login.windows-ppe.net/common',
+      authority: 'https://login.microsoftonline.com/common',
     },
   },
   apiConfig: {
     scopes: ['user.read'],
-    uri: 'https://graph.microsoft-ppe.com/v1.0/me',
+    uri: 'https://graph.microsoft.com/v1.0/me',
   },
 };