+++ lib/msal-browser/CHANGELOG.md
@@ -1566,7 +1566,7 @@ Mon, 07 Aug 2023 18:15:15 GMT
 -   Bundle local version of msal-common into msal-browser #5953 (kshabelko@microsoft.com)
 -   Fix missing idToken in response after refresh #5871 (thomas.norling@microsoft.com)
 -   Improve iframe error detection #5891 (thomas.norling@microsoft.com)
--   Fix: dSTS Token dummy aud claim value for requests with scope input by using v2.0 endpoint (kapjain@microsoft.com)
+-   Fix: Legacy authority token dummy aud claim value for requests with scope input by using v2.0 endpoint (kapjain@microsoft.com)
 -   Exception is thrown in acquireTokenByClientCredential if tenantId is missing #5805 (rginsburg@microsoft.com)
 -   `removeAccount` does not throw if account does not exist in cache #5911 (thomas.norling@microsoft.com)
 -   Remove unused enum (thomas.norling@microsoft.com)